requirements.txt
flask==2.3.3
Werkzeug==2.3.7
google-generativeai>=0.3.2
python-dotenv==0.19.0
Pillow==10.0.0
